pyproject.toml
[tool.poetry]
name = "service_provider_api"
version = "0.1.0"
description = ""
authors = ["Kieran Lavelle"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
pydantic = "^1.10.2"
orjson = "^3.8.2"
boto3 = "^1.26.17"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pytest = "^7.2.0"
moto = "^4.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
